Update blister.py
Fixed bug in which two ways of checking the timestep could clash. If the error caused by a large timestep resulted in h becoming negative at the same time that the regular check of error size was run, the check would not work properly. This might result in the check erroneously increasing dt, and/or in the recorded time t being dt higher that the true time simulated.  I am not aware of this having ocurred in any of my own runs but have fixed it for future use.